refactored the shit outta app.js. Move all the navigation stuff into it's own file
